reorganize viz code
